pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "pyrit-mcp"
version = "0.1.0"
description = "MCP server wrapping Microsoft PyRIT for LLM-orchestrated AI red-teaming"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "ncasfl" }]
keywords = ["mcp", "pyrit", "red-team", "llm", "ai-safety", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mcp>=1.0.0",
    "pyrit>=0.5.0",
    "duckdb>=0.10.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "openai>=1.40.0",
    "azure-ai-inference>=1.0.0b4",
    "huggingface-hub>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/ncasfl/pyrit-mcp"
Repository = "https://github.com/ncasfl/pyrit-mcp"
Issues = "https://github.com/ncasfl/pyrit-mcp/issues"
Changelog = "https://github.com/ncasfl/pyrit-mcp/blob/main/CHANGELOG.md"

[project.scripts]
pyrit-mcp = "pyrit_mcp.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pyrit_mcp*"]

# ── ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py310"
line-length = 100
src = ["pyrit_mcp", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "RUF",  # ruff-specific
]
ignore = [
    "E501",   # line too long — handled by formatter
    "B008",   # do not perform function calls in default arguments
]

[tool.ruff.lint.isort]
known-first-party = ["pyrit_mcp"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ── mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["tests/", "scripts/"]

# ── pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--tb=short -q"
markers = [
    "unit: fast unit tests with no external dependencies",
    "integration: tests requiring live backends (skip in CI unless enabled)",
]

# ── coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["pyrit_mcp"]
omit = ["tests/*", "pyrit_mcp/__init__.py", "pyrit_mcp/server.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true
